pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "proxmox-mcp"
version = "0.1.0"
description = "MCP server for full Proxmox VE management"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Nick Holmquist" }
]
keywords = ["mcp", "proxmox", "virtualization", "claude"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]
dependencies = [
    "mcp>=1.0.0",
    "proxmoxer>=2.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0"
]

[project.scripts]
proxmox-mcp = "proxmox_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/proxmox_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"
